requirements.txt
google-generativeai>=0.8.0
openai>=1.0
pytest>=8.0
python-dotenv>=1.0
pyyaml>=6.0
